Preserve remote compaction behavior
