+++ codex-rs/core/src/compact_remote_v2_attempt.rs
@@ -11,7 +11,6 @@ use crate::responses_metadata::CompactionTurnMetadata;
 use crate::session::session::Session;
 use crate::session::step_context::StepContext;
 use crate::session::turn::built_tools;
-use crate::session::turn_context::TurnContext;
 use codex_protocol::error::Result as CodexResult;
 use codex_protocol::models::ResponseItem;
 use codex_protocol::protocol::TokenUsage;
@@ -20,30 +19,23 @@ use tokio_util::sync::CancellationToken;
 use tracing::info;
 
 pub(super) struct RemoteCompactV2Attempt {
-    pub(super) turn_context: Arc<TurnContext>,
     pub(super) trace_input_history: Vec<ResponseItem>,
     pub(super) prompt_input: Vec<ResponseItem>,
     pub(super) compaction_output: ResponseItem,
     pub(super) token_usage: Option<TokenUsage>,
-    pub(super) compaction_trace: CompactionTraceContext,
+    /// Keeps a session created for standalone compaction alive through lifecycle completion.
+    pub(super) owned_client_session: Option<ModelClientSession>,
 }
 
 pub(super) async fn run_remote_compact_v2_attempt(
     sess: &Arc<Session>,
     step_context: &Arc<StepContext>,
-    client_session: &mut ModelClientSession,
-    compaction_item_id: &str,
+    client_session: Option<&mut ModelClientSession>,
+    compaction_trace: &CompactionTraceContext,
     compaction_metadata: CompactionTurnMetadata,
-    active_context_tokens_before: Option<i64>,
     analytics_details: &mut CompactionAnalyticsDetails,
 ) -> CodexResult<RemoteCompactV2Attempt> {
-    let turn_context = Arc::clone(&step_context.turn);
-    let compaction_trace = sess.services.rollout_thread_trace.compaction_trace_context(
-        turn_context.sub_id.as_str(),
-        compaction_item_id,
-        turn_context.model_info.slug.as_str(),
-        turn_context.provider.info().name.as_str(),
-    );
+    let turn_context = &step_context.turn;
     let mut history = sess.clone_history().await;
     let base_instructions = sess.get_base_instructions().await;
     let (rewritten_outputs, estimated_deleted_tokens) =
@@ -59,13 +51,13 @@ pub(super) async fn run_remote_compact_v2_attempt(
             "rewrote history outputs before remote compaction v2"
         );
     }
-    analytics_details.active_context_tokens_before = active_context_tokens_before;
     if estimated_deleted_tokens > 0 {
         let max_local_deleted_tokens = sess
             .estimated_tokens_after_last_model_generated_item()
             .await;
-        analytics_details.active_context_tokens_before =
-            active_context_tokens_before.map(|active_context_tokens_before| {
+        analytics_details.active_context_tokens_before = analytics_details
+            .active_context_tokens_before
+            .map(|active_context_tokens_before| {
                 active_context_tokens_before
                     .saturating_sub(estimated_deleted_tokens.min(max_local_deleted_tokens))
             });
@@ -102,6 +94,11 @@ pub(super) async fn run_remote_compact_v2_attempt(
         "input": &prompt.input,
         "parallel_tool_calls": prompt.parallel_tool_calls,
     }));
+    let mut owned_client_session = None;
+    let client_session = match client_session {
+        Some(client_session) => client_session,
+        None => owned_client_session.insert(sess.services.model_client.new_session()),
+    };
     let compaction_output_result = run_remote_compaction_request_v2(
         sess,
         turn_context.as_ref(),
@@ -120,11 +117,10 @@ pub(super) async fn run_remote_compact_v2_attempt(
         token_usage,
     } = compaction_output_result?;
     Ok(RemoteCompactV2Attempt {
-        turn_context,
         trace_input_history,
         prompt_input,
         compaction_output,
         token_usage,
-        compaction_trace,
+        owned_client_session,
     })
 }
